fix(richTextEditor) : worked on the review to fix the proplems in the editor
