+++ src/pages/index.tsx
@@ -1,9 +1,9 @@
 
-import Tiptap from '../components/Tiptap'
+import RichTextEditor from '../components/RichTextEditor'
 
 export default function Home() {
     return (
-         <Tiptap />
+         <RichTextEditor />
     )
 }
